requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
cachecontrol==0.12.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
cachetools==4.2.4; python_version ~= '3.5'
certifi==2021.10.8
charset-normalizer==2.0.6; python_version >= '3'
click==8.0.2; python_version >= '3.6'
firebase-admin==5.0.3
flask==2.0.2
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-api-core[grpc]==2.1.0; python_version >= '3.6' and platform_python_implementation != 'PyPy'
google-api-python-client==2.24.0; python_version >= '3.6'
google-auth-httplib2==0.1.0
google-auth==2.3.0; python_version >= '3.6'
google-cloud-core==2.1.0; python_version >= '3.6'
google-cloud-firestore==2.3.4; platform_python_implementation != 'PyPy'
google-cloud-storage==1.42.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-crc32c==1.3.0; python_version >= '3.6'
google-resumable-media==2.0.3; python_version >= '3.6'
googleapis-common-protos==1.53.0; python_version >= '3.6'
grpcio==1.41.0
gunicorn==20.1.0
httplib2==0.20.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.2; python_version >= '3'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.2; python_version >= '3.6'
line-bot-sdk==1.17.0
markupsafe==2.0.1; python_version >= '3.6'
msgpack==1.0.2
packaging==21.0; python_version >= '3.6'
proto-plus==1.19.4; python_version >= '3.6'
protobuf==3.18.1; python_version >= '3.6'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.19.1
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rsa==4.7.2; python_version >= '3.5' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.2; python_version >= '3.6'
